api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
colorama==0.4.4
cycler==0.11.0
datefinder==0.7.1
Deprecated==1.2.13
fastapi==0.73.0
fonttools==4.29.1
h11==0.13.0
hana-ml==2.11.22010700
hdbcli==2.11.14
htmlmin==0.1.12
idna==3.3
importlib-metadata==4.10.1
kiwisolver==1.3.2
loguru==0.6.0
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pep517==0.12.0
Pillow==9.0.0
pydantic==1.9.0
pydot==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
schedule==1.1.0
Shapely==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.0
tqdm==4.62.3
typing_extensions==4.0.1
uvicorn==0.17.1
win32-setctime==1.1.0
wrapt==1.13.3
zipp==3.7.0